Experiences section: - Added experiences accordion - Moved the experiences, skills and technologies content in a store directory
