+++ package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@nextui-org/accordion": "^2.2.5",
     "@nextui-org/button": "^2.0.38",
     "@nextui-org/drawer": "^2.2.5",
     "@nextui-org/react": "^2.6.8",
