+++ src/pages/home/page.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import Hero from "@/pages/home/components/hero/Hero";
 import Skills from "@/pages/home/components/skills/Skills";
 import Technologies from "@/pages/home/components/technologies/Technologies";
+import Experience from "@/pages/home/components/experience/Experience";
 
 const HomePage: React.FC = () => {
   return (
@@ -9,6 +10,7 @@ const HomePage: React.FC = () => {
       <Hero />
       <Skills />
       <Technologies />
+      <Experience />
     </div>
   );
 };
