+++ src/store/experiences.tsx
@@ -1,0 +1,41 @@
+export const experiences = [
+  {
+    company: "AI Law",
+    position: "Frontend Developer",
+    start: "Jul 2024",
+    end: "Sep 2024",
+    description: `Developed multiple frontend features that enable both generative and chatting user interactions with an AI model used for legal matters.
+          Collaborated closely with the backend development team to ensure seamless API integration and proper WebSocket handling for real-time features.
+          Improved the overall responsiveness of the app.
+          Developed an interactive onboarding tutorial, improving user adoption and engagement with the platform.
+          Tackled various bugs and issues and focused on delivering high-quality, well-optimized front-end code.`,
+    technologies: ["React.js", "Next.js", "Typescript"],
+    skills: [
+      "Frontend Architecture",
+      "Data Visualization",
+      "React Optimization",
+      "Git & Version Control",
+      "UI/UX & Responsiveness",
+      "Real-Time Features",
+    ],
+  },
+  {
+    company: "Netcetera",
+    position: "Software Engineer Intern",
+    start: "April 2024",
+    end: "Jun 2024",
+    description: `Worked on the Tech Radar v2 project - an app that displays the company's technology stack on a radar-like UI.
+          Worked on building a fully customizable technology radar chart, with full freedom for future modifications.
+          Developed other UI components to display more details about the technologies.
+          Implemented various optimizations of both the radar and the UI components for better overall performance.
+          Mapped certain details from the backend to the frontend using Java and Spring.`,
+    technologies: ["React.js", "d3.js", "Typescript", "Java", "Spring Boot"],
+    skills: [
+      "Frontend Architecture",
+      "Data Visualization",
+      "React Optimization",
+      "Backend Development",
+      "Git & Version Control",
+    ],
+  },
+];
